requirements.txt
PyYAML==5.3.1
pexpect==4.8.0
pytest==6.0.2
